compiles on windows now with vc 2005 express
git-svn-id: svn://svn.code.sf.net/p/tivodecode/code/tivodecode/trunk@9 3cae7af1-9334-450b-b769-427e0023cdf8

+++ hexlib.c
@@ -10,7 +10,7 @@
 int	nerrors;
 
 static char	*hex = "0123456789abcdef";
-#define HEX(c) (strchr(hex, (c)) - hex)
+#define HEX(c) ((int)(strchr(hex, (c)) - hex))
 
 int
 hexprint(const char *s, unsigned char *p, int n)
